dbt_youtube_dag/requirements.txt
astronomer-cosmos
apache-airflow-providers-snowflake
dbt-snowflake
dbt-core
pandas
google-api-python-client
python-dotenv
snowflake-connector-python 